+++ src/components/Review.css
@@ -1,12 +1,3 @@
-/*className 를 고를땐 앞에 . 를 찍고*/
-/*.center{*/
-/*  min-width: 500px;*/
-/*  margin : 0 auto;*/
-/*  padding: 20px;*/
-/*}*/
-
-/*<body> 태그를 고를땐 그냥 body*/
-/* 기본 세팅 : body 0으로 해주고 시작 */
 .review-bodyBox{
   margin: 0;
   padding: 8%;
@@ -33,8 +24,6 @@
   align-items: center;
   padding: 15px;
 }
-
-/**/
 
 /* Review.css */
 
@@ -90,8 +79,14 @@
 }
 
 .review-rating img {
-  width: 30px;
-  height: 30px;
+  width: 15px;
+  height: 15px;
+  margin-right: 10px;
+}
+
+.review-rating img {
+  width: 15px;
+  height: 15px;
 }
 
 .review-count img {
